Added JWT Authorization for Goals

+++ backend/controllers/goalController.js
@@ -1,5 +1,6 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc  Set goal
 // @route  POST /api/goals
@@ -11,6 +12,7 @@ const setGoal = asyncHandler(async (req, res) => {
     }
 
     const goal = await Goal.create({
+        user: req.user.id,
         text: req.body.text
     });
 
@@ -21,7 +23,7 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route  GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
 
     res.status(200).json(goals);
 });
@@ -32,10 +34,25 @@ const getGoals = asyncHandler(async (req, res) => {
 const updateGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
 
+    // Validate goal id
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found");
     }
+
+    const user = await User.findById(req.user.id);
+
+    // Check for user
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+    
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
+    } 
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, 
         req.body, 
@@ -52,9 +69,24 @@ const updateGoal = asyncHandler(async (req, res) => {
 const deleteGoal = asyncHandler(async (req, res) => {
     const goal = await Goal.findById(req.params.id);
 
+    // Validate goal id
     if (!goal) {
         res.status(400);
         throw new Error("Goal not found");
+    }
+
+    const user = await User.findById(req.user.id);
+
+    // Check for user
+    if (!user) {
+        res.status(401);
+        throw new Error("User not found");
+    }
+    
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error("User not authorized");
     }
 
     await goal.deleteOne();
